cmd/internal/obj/riscv: add assembly support for Zfa extension (#45)
* cmd/internal/obj/riscv: add assembly support of Zfa extension

This patch adds the assembly support of Zfa extension, which is
defined in chapter 26 of RISC-V ISA manual.

Zfa adds several additional floating-point instructions and is
mandatory in RVA23 profile.

Change-Id: I098e4fffb3ba8789453c3c007f58527792104885

* Address some of Joel Sing’s review comments

* convert MOVF/MOVD to FLIS/FLID if appropriate

* support MOVF with floating point constants

* cmd/internal/obj/riscv: improve handling of float point moves

Translate moves from an integer register to a floating point register, or
from a floating point register to an integer register, to the appropriate
move instruction (i.e. FMVXW/FMVWX/FMVXD/FMVDX).

Add support for MOVF with a constant - we previously added support for MOVD
but not for MOVF. Add special handling for 0.0, which we can translate to
a move from the zero register to a floating point register (leveraging the
above mentioned change).

Change-Id: If8df2f5610e69b4ec0af85efb884951024685f5b
Reviewed-on: https://go-review.googlesource.com/c/go/+/703216
Reviewed-by: Meng Zhuo <mengzhuo1203@gmail.com>
Reviewed-by: Mark Freeman <markfreeman@google.com>
Reviewed-by: Michael Knyszek <mknyszek@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Mark Ryan <markdryan@rivosinc.com>

* merge Joel Sing's change

---------

Co-authored-by: Pengcheng Wang <wangpengcheng.pp@bytedance.com>
Co-authored-by: Joel Sing <joel@sing.id.au>