cmd/internal/obj/riscv: implement better bit pattern encoding (#74)
cmd/internal/obj/riscv: implement better bit pattern encoding

add the extractBitAndShift function with an encodeBitPattern function.
This allows the caller to specify a slice of bits that are to be extracted
and encoded, rather than making multiple function calls and combining the
results.

Change-Id: I3d51caa10ecf714f2ad2fb66d38376202c4e0628
Reviewed-on: https://go-review.googlesource.com/c/go/+/702397
Reviewed-by: Junyang Shao <shaojunyang@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Mark Ryan <markdryan@rivosinc.com>
Reviewed-by: Meng Zhuo <mengzhuo1203@gmail.com>
Reviewed-by: Cherry Mui <cherryyz@google.com>

The differnce between this PR and the origin CL is that this PR keep the extractBitAndShift function because Zimop uses it, while the origin CL delete it.

Co-authored-by: Joel Sing <joel@sing.id.au>